requirements.txt
alcohol==0.5
alembic==1.4.3
aniso8601==8.0.0
apispec==4.0.0
appdirs==1.4.4
attrs==20.3.0
autoenv==1.0.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
distlib==0.3.1
filelock==3.0.12
flasgger==0.9.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-rest-paginate==1.1.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mistune==0.8.4
passlib==1.7.4
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
sqlalchemy-json==0.4.0
SQLAlchemy-Utils==0.36.8
text-unidecode==1.3
virtualenv==20.2.1
Werkzeug==0.16.1
zipp==3.4.0
